// components/checkboxes/checkboxesTargets.enum.ts
export enum CheckboxesTargets {
  root = 'root',
  readOnly = 'readOnly',
  label = 'label',
  subLabel = 'subLabel',
  helperText = 'helperText',
  tooltipButton = 'tooltipButton',
  tooltipPopover = 'tooltipPopover',
  checkboxColumnsWrapper = 'checkboxColumnsWrapper',
  checkboxColumns = 'checkboxColumns',
  checkboxWrapper = 'checkboxWrapper',
  checkboxLabelWrapper = 'checkboxLabelWrapper',
  checkboxLabel = 'checkboxLabel',
  checkboxLabelText = 'checkboxLabelText',
  checkbox = 'checkbox',
  errorsRoot = 'errorsRoot',
  errorsUnorderedList = 'errorsUnorderedList',
  errorsUnorderedListItem = 'errorsUnorderedListItem',
  errorsFieldError = 'errorsFieldError',
}